# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Apis",
        "description": "A simplest boilerplate for Laravel Apis",
        "contact": {
            "name": "Gilberto Junior",
            "email": "contact@codehappy.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/permissions": {
            "get": {
                "tags": [
                    "ACL"
                ],
                "summary": "Get permission list.",
                "description": "Paginate all permissions.",
                "operationId": "paginatePermission",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Search by permission name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "-name",
                                "created_at",
                                "-created_at",
                                "updated_at",
                                "-updated_at"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Get permissions by page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Show x permissions per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#components/schemas/Permission"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "pagination": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#components/schemas/Pagination"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
...
        "/auth/me/password": {
            "put": {
                "tags": [
                    "Logged User"
                ],
                "summary": "Change the user's password.",
                "description": "Need to stay logged in to change the password.",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Password changed successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "email": [
                                            "The email field is required."
                                        ],
                                        "password": [
                                            "The selected password is invalid."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Logged User"
                ],
                "summary": "Get the user's profile.",
                "description": "Display the user's profile.",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "put": {
                "tags": [
                    "Logged User"
                ],
                "summary": "Edit the user's profile.",
                "description": "Need to stay logged in to update the user's profile.",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "email": [
                                            "The email must be a valid email address."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in.",
                "description": "Registered users are able to log in.",
                "operationId": "login",
                "requestBody": {
                    "description": "Form request.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Invalid credentials."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "email": [
                                            "The email field is required."
                                        ],
                                        "password": [
                                            "The selected email is invalid."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Permission": {
                "description": "The schema of the permission.",
                "properties": {
                    "id": {
                        "description": "The UUID of the role.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the role.",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "The date of the creation.",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "The date of the last update.",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object",
                "example": {
                    "id": "51a0f12a-5e73-4c04-a5aa-f9322e61e2b7",
                    "name": "permission.viewAny",
                    "created_at": "2022-03-18 20:00:00",
                    "updated_at": "2022-03-18 20:00:00"
                }
            },
            "Role": {
                "description": "The schema of the role.",
                "properties": {
                    "id": {
                        "description": "The UUID of the role.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the role.",
                        "type": "string"
                    },
                    "permissions": {
                        "description": "The permissions of the role.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    },
                    "created_at": {
                        "description": "The date of the creation.",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "The date of the last update.",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object",
                "example": {
                    "id": "75a0f2d3-a0d3-4b0b-b02f-65e70656361d",
                    "name": "Administrator",
                    "permissions": [
                        "permissions.viewAny",
                        "roles.viewAny",
                        "roles.view",
                        "roles.create",
                        "roles.update",
                        "roles.delete",
                        "users.viewAny",
                        "users.view",
                        "users.create",
                        "users.update",
                        "users.delete"
                    ],
                    "created_at": "2022-03-18 20:00:00",
                    "updated_at": "2022-03-18 20:00:00"
                }
            },
...
            "ChangePasswordRequest": {
                "description": "Request data for changing user's password.",
                "required": [
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "password": {
                        "description": "The new password of the user.",
                        "type": "string"
                    },
                    "password_confirmation": {
                        "description": "The password confirmation of the user.",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "password": "secret1234",
                    "password_confirmation": "secret1234"
                }
            },
            "EditProfileRequest": {
                "description": "Request data for updating user's profile.",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "description": "The new name of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The new email of the user.",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object",
                "example": {
                    "name": "Gilberto Junior",
                    "email": "olamundo@gmail.com"
                }
            },
            "LoginRequest": {
                "description": "Request data to generate a valid token to authenticate.",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The email of the user.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password of the user.",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "email": "olamundo@gmail.com",
                    "password": "secret1234"
                }
            },
            "Token": {
                "description": "The response when the token was generated.",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "message": {
                        "description": "The message from response.",
                        "type": "string"
                    },
                    "data": {
                        "description": "The token data.",
                        "properties": {
                            "access_token": {
                                "description": "The new JWT token.",
                                "type": "string"
                            },
                            "bearer_type": {
                                "description": "The type of header authorization.",
                                "type": "string"
                            },
                            "expires_in": {
                                "description": "The time of token expiration in seconds.",
                                "type": "integer",
                                "format": "int32"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Token generated.",
                    "data": {
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9hcGkudGVuYW50MS5oaXJlZDM2NS5kb2NrZXIuaW50ZXJuYWxcL2xvZ2luIiwiaWF0IjoxNjAzNDc2NTE5LCJleHAiOjE2MDM0ODAxMTksIm5iZiI6MTYwMzQ3NjUxOSwianRpIjoiTlNkUXI0M2gzQTFLMWJRSiIsInN1YiI6MSwicHJ2IjoiMWM5OGY5ZmNlMjZkM2JhNDg0NGNkYmQ5M2MzNjcwZTViNGJkYzlmOSJ9.fPbaYuFhOMdrJdMGyGO01_02BzeAUXMO0gHFyDqPcY4",
                        "token_type": "Bearer",
                        "expires_in": 3600
                    }
                }
            },
            "JsonResponse": {
                "description": "The structure of the api response.",
                "properties": {
                    "message": {
                        "description": "The message of the response.",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The error messages of the response if exists.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "The schema of the user.",
                "properties": {
                    "id": {
                        "description": "The UUID of the user.",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the user.",
                        "type": "string"
                    },
                    "roles": {
                        "description": "The roles of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "permissions": {
                        "description": "The permissions of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    },
                    "created_at": {
                        "description": "The date of the creation.",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "The date of the last update.",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object",
                "example": {
                    "id": "beceedca-46b4-4067-a636-eb9dd5ac5569",
                    "name": "John Doe",
                    "email": "john@example.com",
                    "roles": [
                        {
                            "id": "75a0f2d3-a0d3-4b0b-b02f-65e70656361d",
                            "name": "Administrator"
                        }
                    ],
                    "permissions": [
                        {
                            "id": "aa6fa2a0-65a9-4ea6-84ff-7a82faa10439",
                            "name": "custom.permission1"
                        },
                        {
                            "id": "a455ac30-f00d-4a23-a0f9-315e03d82f04",
                            "name": "custom.permission2"
                        }
                    ],
                    "created_at": "2022-03-18 20:00:00",
                    "updated_at": "2022-03-18 20:00:00"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}